split profile into components
